refactor(bench): rename postgres fixtures module to testing
